backend/requirements.txt
# filepath: d:\TASKS\best-card\backend\requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
# Use a wheel-compatible version of pydantic
pydantic==2.3.0
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
google-auth-oauthlib==1.1.0
google-api-python-client==2.107.0
google-auth-httplib2==0.1.1
pyyaml==6.0.1
python-multipart==0.0.6
httpx==0.25.1
PyPDF2==3.0.1
pytest==7.4.3
# Add SSL libraries (choose one of these)
pyopenssl==23.2.0
cryptography==41.0.3
